AT19_CONVERTER/requirements.txt
cryptography==39.0.2
aniso8601==8.0.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.6
coverage==7.1.0
exceptiongroup==1.1.0
Flask==2.0.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.4.3
flask-swagger-ui==4.11.1
googletrans==2.4.0
gunicorn==20.0.4
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==23.0
pluggy==1.0.0
PyJWT==2.6.0
PyMySQL==0.10.1
pytesseract==0.3.10
pytest==7.2.1
pytest-cov==4.0.0
python-dotenv==0.19.0
pytz==2020.1
requests==2.28.1
six==1.15.0
SQLAlchemy==1.3.18
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
flask_cors==3.0.10

